Date assesor should be a function
